+++ CHANGELOG.md
@@ -5,3 +5,18 @@
 ### Added
 
 - Added compatibility with Craft 5.0.0.
+
+### Removed
+
+- Removed the `craft.blitz.getTemplate()` template variable. Use `craft.blitz.includeCached()` or `craft.blitz.includeDynamic()` instead.
+- Removed the `craft.blitz.getUri()` template variable. Use `craft.blitz.fetchUri()` instead.
+- Removed the `blitz/templates/get` controller action.
+- Removed the `cacheElements` config setting. Use `trackElements` instead.
+- Removed the `cacheElementQueries` config setting. Use `trackElementQueries` instead.
+- Removed the `craft.blitz.options.cacheElements()` template variable. Use `craft.blitz.options.trackElements()` instead.
+- Removed the `craft.blitz.options.cacheElementQueries()` template variable. Use `craft.blitz.options.trackElementQueries()` instead.
+- Removed the `createGzipFiles` setting.
+- Removed the `createBrotliFiles` setting.
+- Removed the `BlitzVariable::CACHED_INCLUDE_ACTION` constant. Use `CacheRequestService::CACHED_INCLUDE_ACTION` instead.
+- Removed the `BlitzVariable::DYNAMIC_INCLUDE_ACTION` constant. Use `CacheRequestService::DYNAMIC_INCLUDE_ACTION` instead.
+- Removed the `ElementTypeHelper::LIVE_STATUSES` constant.